Fix staff management and EmployeeDAO Order - AnhND

+++ src/java/dal/OrderDAO.java
@@ -1,6 +1,5 @@
 package dal;
 
-
 import model.Order;
 import model.OrderItem;
 import model.MenuItem;
@@ -19,32 +18,33 @@ public class OrderDAO {
     // =========================================================
     public int createOrder(Order order) {
         String sql = "INSERT INTO `Order` "
-                   + "(customerID, tableID, invoiceID, orderType, tableStatus, "
-                   + " totalAmount, checkoutRequestAt, isStaffConfirmed, "
-                   + " createdAt, orderTime, depositAmount, orderStatus) "
-                   + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+                + "(customerID, tableID, invoiceID, orderType, tableStatus, "
+                + " totalAmount, checkoutRequestAt, isStaffConfirmed, "
+                + " createdAt, orderTime, depositAmount, orderStatus) "
+                + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+                PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
-            ps.setObject    (1,  order.getCustomerID() == 0 ? null : order.getCustomerID());
-            ps.setObject    (2,  order.getTableID()    == 0 ? null : order.getTableID());
-            ps.setObject    (3,  order.getInvoiceID()  == 0 ? null : order.getInvoiceID());
-            ps.setInt       (4,  order.getOrderType());
-            ps.setString    (5,  order.getTableStatus() != null ? order.getTableStatus() : "available");
-            ps.setBigDecimal(6,  order.getTotalAmount());
-            ps.setTimestamp (7,  order.getCheckoutRequestAt());
-            ps.setInt       (8,  order.getIsStaffConfirmed());
-            ps.setTimestamp (9,  order.getCreatedAt() != null
-                                ? order.getCreatedAt()
-                                : new Timestamp(System.currentTimeMillis()));
-            ps.setTimestamp (10, order.getOrderTime());
-            ps.setBigDecimal(11, order.getDepositAmount());
-            ps.setString    (12, order.getOrderStatus() != null ? order.getOrderStatus() : "pending");
+            ps.setObject(1, order.getCustomerID() == 0 ? null : order.getCustomerID());
+            ps.setObject(2, order.getTableID() == 0 ? null : order.getTableID());
+            ps.setObject(3, order.getInvoiceID() == 0 ? null : order.getInvoiceID());
+            ps.setInt(4, order.getOrderType());
+            ps.setString(5, order.getTableStatus() != null ? order.getTableStatus() : "available");
+            ps.setLong(6, order.getTotalAmount());
+            ps.setTimestamp(7, order.getCheckoutRequestAt());
+            ps.setInt(8, order.getIsStaffConfirmed());
+            ps.setTimestamp(9, order.getCreatedAt() != null
+                    ? order.getCreatedAt()
+                    : new Timestamp(System.currentTimeMillis()));
+            ps.setTimestamp(10, order.getOrderTime());
+            ps.setLong(11, order.getDepositAmount());
+            ps.setString(12, order.getOrderStatus() != null ? order.getOrderStatus() : "pending");
 
             ps.executeUpdate();
             ResultSet rs = ps.getGeneratedKeys();
-            if (rs.next()) return rs.getInt(1);
+            if (rs.next())
+                return rs.getInt(1);
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] createOrder lỗi: " + e.getMessage());
@@ -57,17 +57,18 @@ public class OrderDAO {
     // =========================================================
     public Order getActiveOrderByTableId(int tableID) {
         String sql = "SELECT * FROM `Order` "
-                   + "WHERE tableID = ? "
-                   + "  AND tableStatus = 'occupied' "
-                   + "  AND orderStatus NOT IN ('completed', 'checkout') "
-                   + "ORDER BY createdAt DESC LIMIT 1";
+                + "WHERE tableID = ? "
+                + "  AND tableStatus = 'occupied' "
+                + "  AND orderStatus NOT IN ('completed', 'checkout') "
+                + "ORDER BY createdAt DESC LIMIT 1";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, tableID);
             ResultSet rs = ps.executeQuery();
-            if (rs.next()) return mapToOrder(rs);
+            if (rs.next())
+                return mapToOrder(rs);
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] getActiveOrderByTableId lỗi: " + e.getMessage());
@@ -89,16 +90,17 @@ public class OrderDAO {
         String sql = "INSERT INTO OrderItem (orderID, itemID, quantity, note) VALUES (?, ?, ?, ?)";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+                PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
-            ps.setInt   (1, orderID);
-            ps.setInt   (2, itemID);
-            ps.setInt   (3, quantity);
+            ps.setInt(1, orderID);
+            ps.setInt(2, itemID);
+            ps.setInt(3, quantity);
             ps.setString(4, note);
             ps.executeUpdate();
 
             ResultSet rs = ps.getGeneratedKeys();
-            if (rs.next()) return rs.getInt(1);
+            if (rs.next())
+                return rs.getInt(1);
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] addOrderItem lỗi: " + e.getMessage());
@@ -113,7 +115,7 @@ public class OrderDAO {
         String sql = "UPDATE OrderItem SET quantity = ? WHERE orderItemID = ?";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, newQuantity);
             ps.setInt(2, orderItemID);
@@ -132,7 +134,7 @@ public class OrderDAO {
         String sql = "DELETE FROM OrderItem WHERE orderItemID = ?";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, orderItemID);
             return ps.executeUpdate() > 0;
@@ -151,11 +153,12 @@ public class OrderDAO {
         String sql = "SELECT * FROM OrderItem WHERE orderID = ? ORDER BY orderItemID";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, orderID);
             ResultSet rs = ps.executeQuery();
-            while (rs.next()) list.add(mapToOrderItem(rs));
+            while (rs.next())
+                list.add(mapToOrderItem(rs));
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] getOrderItemsByOrderId lỗi: " + e.getMessage());
@@ -170,17 +173,18 @@ public class OrderDAO {
     public List<MenuItem> getMenuItemsByOrderId(int orderID) {
         List<MenuItem> list = new ArrayList<>();
         String sql = "SELECT mi.* "
-                   + "FROM OrderItem oi "
-                   + "JOIN MenuItem mi ON mi.itemID = oi.itemID "
-                   + "WHERE oi.orderID = ? "
-                   + "ORDER BY oi.orderItemID";
+                + "FROM OrderItem oi "
+                + "JOIN MenuItem mi ON mi.itemID = oi.itemID "
+                + "WHERE oi.orderID = ? "
+                + "ORDER BY oi.orderItemID";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, orderID);
             ResultSet rs = ps.executeQuery();
-            while (rs.next()) list.add(mapToMenuItem(rs));
+            while (rs.next())
+                list.add(mapToMenuItem(rs));
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] getMenuItemsByOrderId lỗi: " + e.getMessage());
@@ -195,12 +199,13 @@ public class OrderDAO {
         String sql = "SELECT * FROM OrderItem WHERE orderID = ? AND itemID = ?";
 
         try (Connection conn = getConnection();
-             PreparedStatement ps = conn.prepareStatement(sql)) {
+                PreparedStatement ps = conn.prepareStatement(sql)) {
 
             ps.setInt(1, orderID);
             ps.setInt(2, itemID);
             ResultSet rs = ps.executeQuery();
-            if (rs.next()) return mapToOrderItem(rs);
+            if (rs.next())
+                return mapToOrderItem(rs);
 
         } catch (SQLException e) {
             System.err.println("[OrderDAO] getOrderItemByOrderAndItem lỗi: " + e.getMessage());
@@ -213,20 +218,19 @@ public class OrderDAO {
     // =========================================================
     private Order mapToOrder(ResultSet rs) throws SQLException {
         return new Order(
-            rs.getInt       ("orderID"),
-            rs.getInt       ("customerID"),
-            rs.getInt       ("tableID"),
-            rs.getInt       ("invoiceID"),
-            rs.getString    ("tableStatus"),
-            rs.getBigDecimal("totalAmount"),
-            rs.getTimestamp ("checkoutRequestAt"),
-            rs.getInt       ("isStaffConfirmed"),
-            rs.getTimestamp ("createdAt"),
-            rs.getInt       ("orderType"),
-            rs.getTimestamp ("orderTime"),
-            rs.getBigDecimal("depositAmount"),
-            rs.getString    ("orderStatus")
-        );
+                rs.getInt("orderID"),
+                rs.getInt("customerID"),
+                rs.getInt("tableID"),
+                rs.getInt("invoiceID"),
+                rs.getString("tableStatus"),
+                rs.getLong("totalAmount"),
+                rs.getTimestamp("checkoutRequestAt"),
+                rs.getInt("isStaffConfirmed"),
+                rs.getTimestamp("createdAt"),
+                rs.getInt("orderType"),
+                rs.getTimestamp("orderTime"),
+                rs.getLong("depositAmount"),
+                rs.getString("orderStatus"));
     }
 
     // =========================================================
@@ -234,12 +238,11 @@ public class OrderDAO {
     // =========================================================
     private OrderItem mapToOrderItem(ResultSet rs) throws SQLException {
         return new OrderItem(
-            rs.getInt   ("orderItemID"),
-            rs.getInt   ("orderID"),
-            rs.getInt   ("itemID"),
-            rs.getInt   ("quantity"),
-            rs.getString("note")
-        );
+                rs.getInt("orderItemID"),
+                rs.getInt("orderID"),
+                rs.getInt("itemID"),
+                rs.getInt("quantity"),
+                rs.getString("note"));
     }
 
     // =========================================================
@@ -247,17 +250,16 @@ public class OrderDAO {
     // =========================================================
     private MenuItem mapToMenuItem(ResultSet rs) throws SQLException {
         return new MenuItem(
-            rs.getInt       ("itemID"),
-            rs.getInt       ("categoryID"),
-            rs.getString    ("itemName"),
-            rs.getString    ("description"),
-            rs.getBigDecimal("price"),
-            rs.getBigDecimal("discountPercent"),
-            rs.getBigDecimal("discountedPrice"),
-            rs.getString    ("image"),
-            rs.getInt       ("isAvailable"),
-            rs.getString    ("allergyNotes"),
-            null
-        );
+                rs.getInt("itemID"),
+                rs.getInt("categoryID"),
+                rs.getString("itemName"),
+                rs.getString("description"),
+                rs.getInt("price"),
+                rs.getInt("discountPercent"),
+                rs.getInt("discountedPrice"),
+                rs.getString("image"),
+                rs.getInt("isAvailable"),
+                rs.getString("allergyNotes"),
+                null);
     }
 }
